add widget test cases for calendar link and view + rename calendar view page
