+++ concordia_nav/lib/ui/setting/calendar/calendar_link_view.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 import '../../../data/repositories/calendar.dart';
 import '../../../widgets/custom_appbar.dart';
-import 'calandar_view.dart';
-import 'package:device_calendar/device_calendar.dart';
+import 'calendar_view.dart';
 
 class CalendarLinkView extends StatefulWidget {
   const CalendarLinkView({super.key});
